Move util/helper functions to utils/dir
Consistency across codebase - no reason they need to be in the
bot-commands/ dir while the other utils/helpers are in the utils/ dir.
